app/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pillow==10.4.0
pillow-heif==0.15.0
torch==2.3.1
torchvision==0.18.1
